requirements.txt
Flask
requests
python-dotenv
itsdangerous==2.2.0
email-validator==2.2.0
Flask-Limiter==3.8.0
Flask-Cors==5.0.0   # only if you insist on Live Server (port 5500)
python-dotenv==1.0.1
